Organise files into a module structure.
