+++ meson/compilers.py
@@ -14,10 +14,10 @@
 
 import subprocess, os.path
 import tempfile
-import mesonlib
-import mlog
-from coredata import MesonException
-import coredata
+from .import mesonlib
+from . import mlog
+from .coredata import MesonException
+from . import coredata
 
 """This file contains the data files of all compilers Meson knows
 about. To support a new compiler, add its information below.
