+++ meson/modules/pkgconfig.py
