+++ meson/optinterpreter.py
@@ -12,8 +12,8 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import mparser
-import coredata, mesonlib
+from . import mparser
+from . import coredata, mesonlib
 import os, re
 
 forbidden_option_names = coredata.builtin_options
