+++ meson/wrap.py
@@ -12,12 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-import mlog
+from . import mlog
 import urllib.request, os, hashlib, shutil
 import subprocess
 import sys
 
-import wraptool
+from . import wraptool
 
 class PackageDefinition:
     def __init__(self, fname):
