+++ run_tests.py
@@ -18,15 +18,15 @@ from glob import glob
 import os, subprocess, shutil, sys, signal
 from io import StringIO
 import sys
-import environment
-import mesonlib
-import mlog
-import meson, meson_test, meson_benchmark
+from meson import environment
+from meson import mesonlib
+from meson import mlog
+from meson import meson, meson_test, meson_benchmark
 import argparse
 import xml.etree.ElementTree as ET
 import time
 
-from meson import backendlist
+from meson.meson import backendlist
 
 class TestResult:
     def __init__(self, msg, stdo, stde, conftime=0, buildtime=0, testtime=0):
